+++ backend/src/main/java/com/imane/inventorysystem/dto/SaleResponse.java
@@ -1,12 +1,13 @@
 package com.imane.inventorysystem.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public class SaleResponse {
     private Long id;
     private String transactionId;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private String status;
     private LocalDateTime createdAt;
     private List<SaleItemResponse> items;
@@ -15,8 +16,8 @@ public class SaleResponse {
     public void setId(Long id) { this.id = id; }
     public String getTransactionId() { return transactionId; }
     public void setTransactionId(String transactionId) { this.transactionId = transactionId; }
-    public Double getTotalAmount() { return totalAmount; }
-    public void setTotalAmount(Double totalAmount) { this.totalAmount = totalAmount; }
+    public BigDecimal getTotalAmount() { return totalAmount; }
+    public void setTotalAmount(BigDecimal totalAmount) { this.totalAmount = totalAmount; }
     public String getStatus() { return status; }
     public void setStatus(String status) { this.status = status; }
     public LocalDateTime getCreatedAt() { return createdAt; }
